+++ pkg/test/compute/mock_store.go
@@ -13,7 +13,7 @@ type CallbackStore struct {
 	CreateExecutionFn      func(ctx context.Context, execution store.Execution) error
 	UpdateExecutionStateFn func(ctx context.Context, request store.UpdateExecutionStateRequest) error
 	DeleteExecutionFn      func(ctx context.Context, id string) error
-	GetExecutionCountFn    func(ctx context.Context, state store.ExecutionState) (uint, error)
+	GetExecutionCountFn    func(ctx context.Context, state store.ExecutionState) (uint64, error)
 	CloseFn                func(ctx context.Context) error
 }
 
@@ -41,7 +41,7 @@ func (m *CallbackStore) DeleteExecution(ctx context.Context, id string) error {
 	return m.DeleteExecutionFn(ctx, id)
 }
 
-func (m *CallbackStore) GetExecutionCount(ctx context.Context, state store.ExecutionState) (uint, error) {
+func (m *CallbackStore) GetExecutionCount(ctx context.Context, state store.ExecutionState) (uint64, error) {
 	return m.GetExecutionCountFn(ctx, state)
 }
